Added button bindings for subsystem commands. Additionally added Algae Cleaner commands.
