+++ Big-Bad-Wolf-2025/src/main/java/frc/robot/commands/AlgaeCleaner/AlgaeCleanerShootingCommand.java
@@ -2,33 +2,33 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.AlgaeCleaner;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.constants.AlgaeIntakeConstants;
-import frc.robot.subsystems.AlgaeIntake;
+import frc.robot.constants.AlgaeCleanerConstants;
+import frc.robot.subsystems.AlgaeCleaner;
 
-// Creates a new Algae Intake Stow command.
-public class AlgaeIntakeStowCommand extends Command 
+// Creates a new Algae Cleaner Shooting command.
+public class AlgaeCleanerShootingCommand extends Command 
 {
-  private final AlgaeIntake m_AlgaeIntake;
+  private final AlgaeCleaner m_AlgaeCleaner;
 
   /**
-   * Algae Intake Stow command constructor.
-   * @param algaeIntake Algae Intake subsystem.
+   * Algae Cleaner command constructor.
+   * @param algaeCleaner Algae Cleaner subsystem.
    */
-  public AlgaeIntakeStowCommand(AlgaeIntake algaeIntake) 
+  public AlgaeCleanerShootingCommand(AlgaeCleaner algaeCleaner) 
   {
-    this.m_AlgaeIntake = algaeIntake;
-    addRequirements(m_AlgaeIntake);    
+    this.m_AlgaeCleaner = algaeCleaner;
+    addRequirements(m_AlgaeCleaner);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() 
   {
-    m_AlgaeIntake.setWristPosition(AlgaeIntakeConstants.ALGAE_INTAKE_WRIST_DEFAULT_POSITION);
-    m_AlgaeIntake.setRollerVoltage(AlgaeIntakeConstants.ALGAE_INTAKE_ROLLER_DEFAULT_VOLTAGE);
+    m_AlgaeCleaner.setCleanerWristPosition(AlgaeCleanerConstants.ALGAE_CLEANER_WRIST_SHOOTING_POSITION);
+    m_AlgaeCleaner.setCleanerRollerVoltage(AlgaeCleanerConstants.ALGAE_CLEANER_ROLLER_SHOOTING_VOLTAGE);  
   }
 
   // Called every time the scheduler runs while the command is scheduled.
