+++ Big-Bad-Wolf-2025/src/main/java/frc/robot/commands/AlgaeIntake/AlgaeIntakeStowCommand.java
@@ -2,33 +2,33 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.AlgaeIntake;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.constants.AlgaeIntakeConstants;
 import frc.robot.subsystems.AlgaeIntake;
 
-// Creates a new Algae Intake Processing command.
-public class AlgaeIntakeProcessingCommand extends Command 
+// Creates a new Algae Intake Stow command.
+public class AlgaeIntakeStowCommand extends Command 
 {
   private final AlgaeIntake m_AlgaeIntake;
-  
+
   /**
-   * Algae Intake Processing command constructor.
+   * Algae Intake Stow command constructor.
    * @param algaeIntake Algae Intake subsystem.
    */
-  public AlgaeIntakeProcessingCommand(AlgaeIntake algaeIntake) 
+  public AlgaeIntakeStowCommand(AlgaeIntake algaeIntake) 
   {
     this.m_AlgaeIntake = algaeIntake;
-    addRequirements(m_AlgaeIntake);
+    addRequirements(m_AlgaeIntake);    
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() 
   {
-    m_AlgaeIntake.setWristPosition(AlgaeIntakeConstants.ALGAE_INTAKE_WRIST_PROCESSING_POSITION);
-    m_AlgaeIntake.setRollerVoltage(AlgaeIntakeConstants.ALGAE_INTAKE_ROLLER_PROCESSING_VOLTAGE);
+    m_AlgaeIntake.setWristPosition(AlgaeIntakeConstants.ALGAE_INTAKE_WRIST_DEFAULT_POSITION);
+    m_AlgaeIntake.setRollerVoltage(AlgaeIntakeConstants.ALGAE_INTAKE_ROLLER_DEFAULT_VOLTAGE);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
@@ -40,7 +40,7 @@ public class AlgaeIntakeProcessingCommand extends Command
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted)
+  public void end(boolean interrupted) 
   {
     // Intentionally Empty.
   }
