+++ Big-Bad-Wolf-2025/src/main/java/frc/robot/commands/Elevator/ElevatorLevelTwo.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Elevator;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.constants.ElevatorConstants;
@@ -12,14 +12,14 @@ import frc.robot.subsystems.Elevator;
 /** 
  * You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands 
  */
-public class ElevatorUp extends Command 
+public class ElevatorLevelTwo extends Command 
 {
-  private final Elevator m_Elevator;
+   private final Elevator m_Elevator;
   /** 
-   * Creates a new ElevatorUp. 
+   * Creates a new ElevatorLevelTwo. 
    * @param p_Elevator subsystem created in robotcontainer
    */
-  public ElevatorUp(Elevator p_Elevator)
+  public ElevatorLevelTwo(Elevator p_Elevator) 
   {
     this.m_Elevator = p_Elevator;
     addRequirements(this.m_Elevator);
@@ -31,7 +31,7 @@ public class ElevatorUp extends Command
   @Override
   public void initialize() 
   {
-    m_Elevator.elevatorRequest(Global.MODE.VOLTAGE, ElevatorConstants.ELEVATOR_UP_VOLTAGE);// Applies voltage for elevator to go up
+    m_Elevator.elevatorRequest(Global.MODE.POSITION, ElevatorConstants.ELEVATOR_LEVEL_TWO);// Sets elevator to Level 2
   }
 
   /**
